backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
mysql-connector-python==8.2.0
python-dotenv==1.0.0
PyMySQL==1.1.0
PyJWT==2.8.0
Werkzeug==3.0.1
